+++ mllc/src/typechecker/infer.rs
@@ -940,6 +940,7 @@ impl Checker {
             ty: final_ty,
             clauses: tclauses,
             specialized: false,
+            spec_origin: None,
             dict_params: vec![],
             derived_strict: false,
         };
@@ -2753,6 +2754,7 @@ impl Checker {
                     where_binds: vec![],
                 }],
                 specialized: false,
+                spec_origin: None,
             dict_params: vec![],
             derived_strict: false,
             };
@@ -2860,6 +2862,7 @@ impl Checker {
                 where_binds: vec![],
             }],
             specialized: false,
+            spec_origin: None,
             dict_params: vec![],
             derived_strict: false,
         }